Attempted refactor
